+++ trunk/i_mod_pers_prod.h
@@ -7,11 +7,12 @@
 #define EXT_PPROD extern
 #endif
 
+#include "tabela_codigos.h"
 #include "i_estruturas.h"
 
-EXT_PPROD size_t cadastrar_produto(const struct produto *dev);
-EXT_PPROD size_t consultar_produto(struct produto *dev);
-EXT_PPROD size_t editar_produto(struct produto *dev);
-EXT_PPROD size_t remover_produto(const char *cod);
+EXT_PPROD int cadastrar_produto(const struct produto *dev);
+EXT_PPROD int consultar_produto(struct produto *dev);
+EXT_PPROD int editar_produto(struct produto *dev);
+EXT_PPROD int remover_produto(const char *cod);
 
 #endif
